requirements.txt
# Core news scraping dependencies
lxml>=3.3.5
beautifulsoup4>=4.3.2
# The 0.8.4.1 version is broken. cf. https://github.com/buriy/python-readability/issues/194
readability-lxml==0.8.1
langdetect>=1.0.7
python-dateutil>=2.4.0
dotmap>=1.2.17
newspaper4k>=0.9.3.1
lxml-html-clean>=0.1.1
typing-extensions>=4.7.0

# Email enrichment and validation
python-dotenv
requests
email-validator
dnspython
pandas
tqdm

# Optional database support (if needed)
PyMySQL>=0.7.9
psycopg2-binary>=2.8.4
redis

# Legacy dependencies (kept for compatibility)
hjson>=1.5.8
elasticsearch>=2.4
plac>=0.9.6
PyDispatcher>=2.0.5
warcio>=1.3.3
ago>=0.0.9
six>=1.10.0
hurry.filesize>=0.9
bs4
faust-cchardet>=2.1.18
boto3

# Windows support
pywin32>=220 ; sys_platform == 'win32'
